# src/openApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hoeken's Anchor Alarm API",
    "version": "2.0.0",
    "description": "HTTP API for dropping, adjusting, and raising the anchor managed by the Hoeken's Anchor Alarm plugin.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "/plugins/y2k-anchor-alarm"
    }
  ],
  "tags": [
    {
      "name": "anchor",
      "description": "Anchor control operations"
    },
    {
      "name": "config",
      "description": "Plugin configuration exposed to the UI"
    }
  ],
  "components": {
    "schemas": {
      "Position": {
        "type": "object",
        "required": ["latitude", "longitude"],
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double",
            "minimum": -90,
            "maximum": 90,
            "description": "Latitude in decimal degrees (WGS84)."
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "minimum": -180,
            "maximum": 180,
            "description": "Longitude in decimal degrees (WGS84)."
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": ["statusCode", "state"],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "state": {
            "type": "string",
            "enum": ["COMPLETED"]
          }
        }
      },
      "Config": {
        "type": "object",
        "description": "UI-relevant plugin configuration. Properties are omitted when unset; the UI falls back to its own defaults.",
        "properties": {
          "fleetFilterRadius": {
            "type": "integer",
            "minimum": 0,
            "description": "Radius (meters) around own vessel for displaying other vessels and historical tracks."
          },
          "connectionType": {
            "type": "string",
            "enum": ["POLLING", "WEBSOCKET"],
            "description": "How the UI connects to SignalK for live data updates."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["statusCode", "state", "message"],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 403
          },
          "state": {
            "type": "string",
            "enum": ["FAILED"]
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "AnchorError": {
        "description": "Request rejected due to an anchor validation or state error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "statusCode": 403,
              "state": "FAILED",
              "message": "position with latitude and longitude required"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected internal error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "statusCode": 500,
              "state": "FAILED",
              "message": "internal error"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/dropAnchor": {
      "post": {
        "tags": ["anchor"],
        "summary": "Drop the anchor at a given position",
        "description": "Records the supplied position as the anchor location and (optionally) sets the alarm radius. Starts the anchor watch.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["position"],
                "properties": {
                  "position": {
                    "$ref": "#/components/schemas/Position"
                  },
                  "zone": {
                    "type": "object",
                    "required": ["type"],
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": ["circle", "sector"]
                      },
                      "radius": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0,
                        "description": "Alarm radius in meters. Used by both `circle` and `sector` shapes."
                      },
                      "startAngle": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0,
                        "maximum": 360,
                        "description": "Sector only: clockwise bearing in degrees (0 = true north) of the start of the safe arc."
                      },
                      "endAngle": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0,
                        "maximum": 360,
                        "description": "Sector only: clockwise bearing in degrees (0 = true north) of the end of the safe arc. When `endAngle < startAngle` the arc wraps across 0°/360°."
                      }
                    },
                    "description": "Watch zone configuration. If omitted, the previously configured zone is retained."
                  },
                  "radius": {
                    "type": "number",
                    "format": "double",
                    "minimum": 0,
                    "description": "Legacy: alarm radius in meters. If both `zone` and `radius` are provided, `zone` wins."
                  }
                }
              },
              "example": {
                "position": {
                  "latitude": 37.8199,
                  "longitude": -122.4783
                },
                "zone": {
                  "type": "circle",
                  "radius": 30
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Anchor dropped successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AnchorError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/setZone": {
      "post": {
        "tags": ["anchor"],
        "summary": "Set the anchor watch zone",
        "description": "Updates the watch zone (shape + parameters) for the currently dropped anchor. The anchor position is unchanged. Requires that an anchor has been dropped and that a GPS position is available.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["zone"],
                "properties": {
                  "zone": {
                    "type": "object",
                    "required": ["type"],
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": ["circle", "sector"],
                        "description": "Watch zone shape."
                      },
                      "radius": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0,
                        "description": "Alarm radius in meters. Used by both `circle` and `sector` shapes."
                      },
                      "startAngle": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0,
                        "maximum": 360,
                        "description": "Sector only: clockwise bearing in degrees (0 = true north) of the start of the safe arc."
                      },
                      "endAngle": {
                        "type": "number",
                        "format": "double",
                        "minimum": 0,
                        "maximum": 360,
                        "description": "Sector only: clockwise bearing in degrees (0 = true north) of the end of the safe arc. When `endAngle < startAngle` the arc wraps across 0°/360°."
                      }
                    }
                  }
                }
              },
              "example": {
                "zone": {
                  "type": "circle",
                  "radius": 45
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zone updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AnchorError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/raiseAnchor": {
      "post": {
        "tags": ["anchor"],
        "summary": "Raise the anchor",
        "description": "Clears the anchor position and radius and stops the anchor watch.",
        "responses": {
          "200": {
            "description": "Anchor raised successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AnchorError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/resetWindReference": {
      "post": {
        "tags": ["anchor"],
        "summary": "Reset wind direction reference",
        "description": "Records the current wind direction as the new reference for the wind direction shift alarm. Use after acknowledging a wind shift.",
        "responses": {
          "200": {
            "description": "Wind reference reset successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/ui-config": {
      "get": {
        "tags": ["config"],
        "summary": "Get UI-relevant plugin configuration",
        "description": "Returns plugin configuration values that the web UI needs at startup. Unset properties are omitted; the UI applies its own defaults in that case.",
        "responses": {
          "200": {
            "description": "Plugin configuration returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                },
                "example": {
                  "fleetFilterRadius": 500,
                  "connectionType": "WEBSOCKET"
                }
              }
            }
          }
        }
      }
    }
  }
}